% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scp.R
\name{scp}
\alias{scp}
\alias{scp_download}
\alias{scp_upload}
\title{SCP (Secure Copy)}
\usage{
scp_download(session = ssh_connect(), files, to = ".", verbose = TRUE)

scp_upload(session = ssh_connect(), files, to = ".", verbose = TRUE)
}
\arguments{
\item{session}{ssh connnection created with \code{\link[=ssh_connect]{ssh_connect()}}}

\item{files}{path to files or directory to transfer}

\item{to}{existing directory on the destination where \code{files} will be copied into}

\item{verbose}{print progress while copying files}
}
\description{
Upload and download files to/from the SSH server via the scp protocol.
Directories in the \code{files} argument are automatically trasversed and
uploaded / downloaded recursively.
}
\details{
Default path \code{to = "."} means that files get downloaded to the current
working directory and uploaded to the user home directory on the server.
}
