% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{ssh_connect}
\alias{ssh_connect}
\alias{ssh}
\title{SSH Client}
\usage{
ssh_connect(host = "dev.opencpu.org:22", keyfile = NULL, passwd = askpass)
}
\arguments{
\item{host}{an ssh server string of the form \code{[user@]hostname[:@port]}}

\item{keyfile}{path to private key. If \code{NULL} the default user key is tried.}

\item{passwd}{either a string or a callback function for password prompt}
}
\description{
Create an ssh session using \code{ssh_connect()}. Then use the other functions to
run commands or create a tunnel via this ssh session.
}
\examples{
\dontrun{
ssh_exec(command = c(
  'curl -O https://cran.r-project.org/src/contrib/jsonlite_1.5.tar.gz',
  'R CMD check jsonlite_1.5.tar.gz',
  'rm -f jsonlite_1.5.tar.gz'
))}
}
\seealso{
Other ssh: \code{\link{ssh_exec}}, \code{\link{ssh_tunnel}}
}
\concept{ssh}
