% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scp.R
\name{scp}
\alias{scp}
\alias{scp_read_file}
\title{Upload / Download via SCP}
\usage{
scp_read_file(session = ssh_connect(), path)
}
\arguments{
\item{session}{ssh connnection created with \code{\link[=ssh_connect]{ssh_connect()}}}

\item{path}{path to the file on the server}
}
\description{
Upload and download files and directories to/from the SSH server.
}
