% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scp.R
\name{scp}
\alias{scp}
\alias{scp_download}
\alias{scp_upload}
\title{Upload / Download via SCP}
\usage{
scp_download(session = ssh_connect(), files, to = ".", verbose = TRUE)

scp_upload(session = ssh_connect(), files, to = ".", verbose = TRUE)
}
\arguments{
\item{session}{ssh connnection created with \code{\link[=ssh_connect]{ssh_connect()}}}

\item{files}{path to directory or file to upload or download}

\item{to}{path to an existing location on the local system (for download) or server
system (for upload) where the target \code{files} will be copied into.
downloaded files. Default copies everything into working directory for downloads,
and home directory for uploads.}

\item{verbose}{print file paths and sizes while copying}
}
\description{
Upload / download files and directories to/from the SSH server.
}
