% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/command.R, R/session.R, R/tunnel.R
\name{exec}
\alias{exec}
\alias{ssh}
\alias{tunnel}
\title{SSH tunnel}
\usage{
exec(session = ssh(), command = "whoami")

ssh(host = "dev.opencpu.org:22", keyfile = NULL, passwd = askpass)

tunnel(session = ssh(), port = 5555,
  target = "ds043942.mongolab.com:43942")
}
\arguments{
\item{session}{existing ssh connnection created with \link{ssh}}

\item{command}{The command to execute (e.g. \code{ls ~/ -al | grep -i reports}).}

\item{host}{an ssh server string of the form \code{[user@]hostname[:@port]}}

\item{keyfile}{path to private key. If \code{NULL} the default user key is tried.}

\item{passwd}{either a string or a callback function for password prompt}

\item{port}{integer of local port on which to listen for incoming connections}

\item{target}{string with target host and port to connnect to via ssh tunnel}
}
\description{
Create an SSH tunnel
}
