% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R, R/tunnel.R
\name{ssh}
\alias{ssh}
\alias{tunnel}
\title{SSH tunnel}
\usage{
ssh(host = "dev.opencpu.org:22", passwd = askpass)

tunnel(session = ssh(), port = 5555,
  target = "ds043942.mongolab.com:43942")
}
\arguments{
\item{host}{an ssh server string of the form \code{[user@]hostname[:@port]}}

\item{passwd}{either a string or a callback function for password prompt}

\item{session}{existing ssh connnection created with \link{ssh}}

\item{port}{integer of local port on which to listen for incoming connections}

\item{target}{string with target host and port to connnect to via ssh tunnel}
}
\description{
Create an SSH tunnel
}
