% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/command.R, R/connect.R, R/tunnel.R
\name{ssh_exec}
\alias{ssh_exec}
\alias{ssh_connect}
\alias{ssh}
\alias{ssh_tunnel}
\title{SSH Client}
\usage{
ssh_exec(session = ssh_connect(), command = "whoami", std_out = stdout(),
  std_err = stderr())

ssh_connect(host = "dev.opencpu.org:22", keyfile = NULL, passwd = askpass)

ssh_tunnel(session = ssh_connect(), port = 5555,
  target = "ds043942.mongolab.com:43942")
}
\arguments{
\item{session}{existing ssh connnection created with \link{ssh}}

\item{command}{The command to execute (e.g. \code{ls ~/ -al | grep -i reports}).}

\item{std_out}{callback function, filename, or connection object to handle stdout stream}

\item{std_err}{callback function, filename, or connection object to handle stderr stream}

\item{host}{an ssh server string of the form \code{[user@]hostname[:@port]}}

\item{keyfile}{path to private key. If \code{NULL} the default user key is tried.}

\item{passwd}{either a string or a callback function for password prompt}

\item{port}{integer of local port on which to listen for incoming connections}

\item{target}{string with target host and port to connnect to via ssh tunnel}
}
\description{
Create an ssh session using \code{ssh_connect()}. Then use the other functions to
run commands or create a tunnel via this ssh session.
}
\examples{
\dontrun{
ssh_exec(command = c(
  'curl -O https://cran.r-project.org/src/contrib/jsonlite_1.5.tar.gz',
  'R CMD check jsonlite_1.5.tar.gz',
  'rm -f jsonlite_1.5.tar.gz'
))}
}
