% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tunnel.R
\name{tunnel}
\alias{tunnel}
\title{SSH tunnel}
\usage{
tunnel(ssh = "dev.opencpu.org:22", target = "ds043942.mongolab.com:43942",
  listen = 5555, passwd = askpass)
}
\arguments{
\item{ssh}{an ssh server string of the form `[user@]hostname[:@port]`}

\item{target}{target sever string of the form `hostname[:port]`}

\item{listen}{port on which to listen for incoming connections}

\item{passwd}{string or callback function to supply ssh password}
}
\description{
Create an SSH tunnel
}
