% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R, R/tunnel.R
\name{ssh}
\alias{ssh}
\alias{tunnel}
\title{SSH tunnel}
\usage{
ssh(host = "dev.opencpu.org:22", passwd = askpass)

tunnel(ssh = "dev.opencpu.org:22", target = "ds043942.mongolab.com:43942",
  listen = 5555, passwd = askpass)
}
\arguments{
\item{host}{an ssh server string of the form `[user@]hostname[:@port]`}

\item{passwd}{string or callback function to supply ssh password}

\item{ssh}{an ssh server string of the form `[user@]hostname[:@port]`}

\item{target}{target sever string of the form `hostname[:port]`}

\item{listen}{port on which to listen for incoming connections}
}
\description{
Create an SSH tunnel
}
